pyproject.toml
[project]
name = "coronary-stenosis-detection"
version = "1.0.0"
description = ""
authors = [
    {name = "Kamil Rudny (krudny)"},
    {name = "Piotr Rusak (piotrrusak)"},
    {name = "Łukasz Kwinta (lkwinta)", email = "lukaszmind@gmail.com"}
]
readme = "README.md"

requires-python = "==3.12.*"

dependencies = [
    "numpy",
    "ipykernel",
    "jupyterlab",
    "ipywidgets",
    "pandas",
    "seaborn",
    "datasets",
    "opencv-python",
    "pycocotools",
    "segmentation_models_pytorch",
    "albumentations",
    "scikit-learn",
    "coronary-analysis",
    "matplotlib",
    "optuna",
    "optuna-dashboard",
]

[dependency-groups]
dev = [
    "pre-commit",
    "ruff"
]

[tool.ruff.lint]
ignore = [
    "E731",  # lambda expressions
]

[tool.uv.sources]
coronary-analysis = { path = "coronary-analysis", editable = true }
